+++ src/pages/Models/index.tsx
@@ -14,7 +14,6 @@ import {
   InputLabel,
   FormControl,
   Fab,
-  Tooltip,
   Chip,
   Button,
 } from "@mui/material";
@@ -22,8 +21,6 @@ import TuneIcon from "@mui/icons-material/Tune";
 import WhatsAppIcon from "@mui/icons-material/WhatsApp";
 import modelsData from "../../db/models-details.json";
 import ImageBox from "../../components/ImageBox";
-import useScrollTrigger from "@mui/material/useScrollTrigger";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import Fade from "@mui/material/Fade";
 import Snackbar, { SnackbarOrigin } from "@mui/material/Snackbar";
 import { Model } from "../../types";
@@ -68,48 +65,10 @@ interface Filters {
 interface State extends SnackbarOrigin {
   open: boolean;
 }
-
-interface Props {
-  window?: () => Window;
-  children: React.ReactElement;
-}
-
 const MenuItemStyles = {
   bgcolor: "black",
   color: "gold",
 };
-
-// === === === BEGIN SCROLL TO TOP === === === //
-function ScrollTop(props: Props) {
-  const { children, window } = props;
-  const trigger = useScrollTrigger({
-    target: window ? window() : undefined,
-    disableHysteresis: true,
-    threshold: 100,
-  });
-
-  const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    const anchor = (
-      (event.target as HTMLDivElement).ownerDocument || document
-    ).querySelector("#back-to-top-anchor");
-
-    if (anchor) {
-      anchor.scrollIntoView({
-        behavior: "smooth",
-        block: "center",
-      });
-    }
-  };
-
-  return (
-    <Fade in={trigger}>
-      <Box onClick={handleClick} role="presentation" sx={{ zIndex: 9999 }}>
-        {children}
-      </Box>
-    </Fade>
-  );
-}
-// === === === END SCROLL TO TOP === === === //
 
 export default function Models() {
   const [favorites, setFavorites] = useState<number[]>([]);
@@ -750,27 +709,6 @@ export default function Models() {
         )}
       </Grid>
 
-      <ScrollTop>
-        <Tooltip title="Voltar ao topo" arrow>
-          <Fab
-            size="small"
-            aria-label="scroll back to top"
-            sx={{
-              position: "fixed",
-              bottom: 16,
-              right: 16,
-              zIndex: 9999,
-              opacity: 0.4,
-              transition: "opacity 0.5s ease",
-              "&:hover": {
-                opacity: 1,
-              },
-            }}
-          >
-            <KeyboardArrowUpIcon />
-          </Fab>
-        </Tooltip>
-      </ScrollTop>
     </Box>
   );
 }
