--- src/routes/index-bk2505-0951.tsx
@@ -1,50 +1,0 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Home from "../pages/Home";
-import About from "../pages/About";
-import Register from "../pages/Register";
-import Admin from "../pages/Admin";
-import Login from "../pages/Login";
-import ProtectedRoute from "./ProtectedRoute";
-import AuthProvider from "../providers/AuthProvider";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Contact from "../pages/Contact";
-import Models from "../pages/Models";
-import { Stack } from "@mui/material";
-import ModelProfile from "../pages/Models/ModelProfile";
-import Advertise from "../pages/Advertise";
-import Favorites from "../pages/Favorites";
-import Terms from "../pages/Terms";
-
-export default function AppRouter() {
-  return (
-    <BrowserRouter>
-      <AuthProvider isSignedIn={true}>
-        <Header />
-        <Stack my={10}>
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/about" element={<About />} />
-            <Route path="/acompanhantes" element={<Models />} />
-            <Route path="/acompanhantes/:name" element={<ModelProfile name="" />} />
-            <Route path="/contato" element={<Contact />} />
-            <Route path="/anuncie" element={<Advertise />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/register" element={<Register />} />
-            <Route path="/favoritos" element={<Favorites />} />
-            <Route path="/terms" element={<Terms />} />
-            <Route
-              path="/admin"
-              element={
-                <ProtectedRoute>
-                  <Admin />
-                </ProtectedRoute>
-              }
-            />
-          </Routes>
-        </Stack>
-        <Footer />
-      </AuthProvider>
-    </BrowserRouter>
-  );
-}
